// apps/api/src/domains/decisions/models/interfaces.ts
import type { IDomainNodeProps, ITimestampedDetails } from "../../shared/models/shared.js";

export interface IDecisionDetailsProps extends ITimestampedDetails {
  rationale: string;
  impact: string;
  decidedAt: Date;
  supersedesDecisionNodeId: string | null;
}

export interface IDecisionNodeProps extends Omit<IDomainNodeProps<"decision">, "nodeType"> {
  details: IDecisionDetailsProps | null;
}